Refactored to Railtie and added support for Sprockets 3.0+.
This branch has support for Rails 5 and uses the assets pipeline to compile
`elm` files.
