+++ lib/elm/rails.rb
@@ -1,9 +1,2 @@
-require "elm/rails/engine"
-require "elm/rails/version"
-
-module Elm
-  module Rails
-    mattr_accessor :elm_make_path
-  end
-end
-
+require "elm/rails/helper"
+require "elm/rails/railtie"
