+++ lib/elm/rails/helper.rb
